// packages/fhir-r5/src/v5.0.0/explanationofbenefitdetail1/types.ts
import type {
  Extension,
  Identifier,
  CodeableConcept,
  Money,
  Quantity,
  Element,
} from "../core/types";
import type { ExplanationOfBenefitReviewOutcome } from "../explanationofbenefitreviewoutcome/types";
import type { ExplanationOfBenefitAdjudication } from "../explanationofbenefitadjudication/types";
import type { ExplanationOfBenefitSubDetail1 } from "../explanationofbenefitsubdetail1/types";

/** Generated from FHIR JSON Schema */

/** This resource provides: the claim details; adjudication details from the processing of a Claim; and optionally account balance information, for informing the subscriber of the benefits provided. */
export interface ExplanationOfBenefitDetail1 {
  id?: string;

  extension?: Extension[];

  modifierExtension?: Extension[];

  traceNumber?: Identifier[];

  revenue?: CodeableConcept;

  productOrService?: CodeableConcept;

  productOrServiceEnd?: CodeableConcept;

  modifier?: CodeableConcept[];

  patientPaid?: Money;

  quantity?: Quantity;

  unitPrice?: Money;

  factor?: number;

  _factor?: Element;

  tax?: Money;

  net?: Money;

  noteNumber?: number[];

  _noteNumber?: Element[];

  reviewOutcome?: ExplanationOfBenefitReviewOutcome;

  adjudication?: ExplanationOfBenefitAdjudication[];

  subDetail?: ExplanationOfBenefitSubDetail1[];
}